for #2567, move selectItems related method from ShardingSelectOptimizedStatement to SelectItems
