+++ sharding-core/sharding-core-merge/src/main/java/org/apache/shardingsphere/core/merge/dql/groupby/GroupByMemoryMergedResult.java
@@ -79,7 +79,7 @@ public final class GroupByMemoryMergedResult extends MemoryMergedResult {
             dataMap.put(groupByValue, new MemoryQueryResultRow(queryResult));
         }
         if (!aggregationMap.containsKey(groupByValue)) {
-            Map<AggregationSelectItem, AggregationUnit> map = Maps.toMap(optimizedStatement.getAggregationSelectItems(), new Function<AggregationSelectItem, AggregationUnit>() {
+            Map<AggregationSelectItem, AggregationUnit> map = Maps.toMap(optimizedStatement.getSelectItems().getAggregationSelectItems(), new Function<AggregationSelectItem, AggregationUnit>() {
                 
                 @Override
                 public AggregationUnit apply(final AggregationSelectItem input) {
@@ -91,7 +91,7 @@ public final class GroupByMemoryMergedResult extends MemoryMergedResult {
     }
     
     private void aggregate(final QueryResult queryResult, final GroupByValue groupByValue, final Map<GroupByValue, Map<AggregationSelectItem, AggregationUnit>> aggregationMap) throws SQLException {
-        for (AggregationSelectItem each : optimizedStatement.getAggregationSelectItems()) {
+        for (AggregationSelectItem each : optimizedStatement.getSelectItems().getAggregationSelectItems()) {
             List<Comparable<?>> values = new ArrayList<>(2);
             if (each.getDerivedAggregationItems().isEmpty()) {
                 values.add(getAggregationValue(queryResult, each));
@@ -112,7 +112,7 @@ public final class GroupByMemoryMergedResult extends MemoryMergedResult {
     
     private void setAggregationValueToMemoryRow(final Map<GroupByValue, MemoryQueryResultRow> dataMap, final Map<GroupByValue, Map<AggregationSelectItem, AggregationUnit>> aggregationMap) {
         for (Entry<GroupByValue, MemoryQueryResultRow> entry : dataMap.entrySet()) {
-            for (AggregationSelectItem each : optimizedStatement.getAggregationSelectItems()) {
+            for (AggregationSelectItem each : optimizedStatement.getSelectItems().getAggregationSelectItems()) {
                 entry.getValue().setCell(each.getIndex(), aggregationMap.get(entry.getKey()).get(each).getResult());
             }
         }
