+++ sharding-core/sharding-core-merge/src/main/java/org/apache/shardingsphere/core/merge/dql/groupby/GroupByStreamMergedResult.java
@@ -77,13 +77,14 @@ public final class GroupByStreamMergedResult extends OrderByStreamMergedResult {
     
     private boolean aggregateCurrentGroupByRowAndNext() throws SQLException {
         boolean result = false;
-        Map<AggregationSelectItem, AggregationUnit> aggregationUnitMap = Maps.toMap(optimizedStatement.getAggregationSelectItems(), new Function<AggregationSelectItem, AggregationUnit>() {
-            
-            @Override
-            public AggregationUnit apply(final AggregationSelectItem input) {
-                return AggregationUnitFactory.create(input.getType());
-            }
-        });
+        Map<AggregationSelectItem, AggregationUnit> aggregationUnitMap = Maps.toMap(
+                optimizedStatement.getSelectItems().getAggregationSelectItems(), new Function<AggregationSelectItem, AggregationUnit>() {
+                    
+                    @Override
+                    public AggregationUnit apply(final AggregationSelectItem input) {
+                        return AggregationUnitFactory.create(input.getType());
+                    }
+                });
         while (currentGroupByValues.equals(new GroupByValue(getCurrentQueryResult(), optimizedStatement.getGroupBy().getItems()).getGroupValues())) {
             aggregate(aggregationUnitMap);
             cacheCurrentRow();
