+++ sharding-core/sharding-core-optimize/src/main/java/org/apache/shardingsphere/core/optimize/statement/sharding/dml/select/item/SelectItems.java
@@ -21,10 +21,15 @@ import com.google.common.base.Optional;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.Setter;
+import org.apache.shardingsphere.core.parse.sql.context.selectitem.AggregationDistinctSelectItem;
+import org.apache.shardingsphere.core.parse.sql.context.selectitem.AggregationSelectItem;
+import org.apache.shardingsphere.core.parse.sql.context.selectitem.DistinctSelectItem;
 import org.apache.shardingsphere.core.parse.sql.context.selectitem.SelectItem;
 
 import java.util.Collection;
 import java.util.LinkedHashSet;
+import java.util.LinkedList;
+import java.util.List;
 
 /**
  * Select items.
@@ -55,5 +60,51 @@ public final class SelectItems {
             }
         }
         return Optional.absent();
+    }
+    
+    /**
+     * Get aggregation select items.
+     *
+     * @return aggregation select items
+     */
+    public List<AggregationSelectItem> getAggregationSelectItems() {
+        List<AggregationSelectItem> result = new LinkedList<>();
+        for (SelectItem each : items) {
+            if (each instanceof AggregationSelectItem) {
+                AggregationSelectItem aggregationSelectItem = (AggregationSelectItem) each;
+                result.add(aggregationSelectItem);
+                result.addAll(aggregationSelectItem.getDerivedAggregationItems());
+            }
+        }
+        return result;
+    }
+    
+    /**
+     * Get distinct select items.
+     *
+     * @return distinct select items
+     */
+    public Optional<DistinctSelectItem> getDistinctSelectItem() {
+        for (SelectItem each : items) {
+            if (each instanceof DistinctSelectItem) {
+                return Optional.of((DistinctSelectItem) each);
+            }
+        }
+        return Optional.absent();
+    }
+    
+    /**
+     * Get aggregation distinct select items.
+     *
+     * @return aggregation distinct select items
+     */
+    public List<AggregationDistinctSelectItem> getAggregationDistinctSelectItems() {
+        List<AggregationDistinctSelectItem> result = new LinkedList<>();
+        for (SelectItem each : items) {
+            if (each instanceof AggregationDistinctSelectItem) {
+                result.add((AggregationDistinctSelectItem) each);
+            }
+        }
+        return result;
     }
 }
