+++ sharding-core/sharding-core-optimize/src/main/java/org/apache/shardingsphere/core/optimize/statement/sharding/dml/select/pagination/Pagination.java
@@ -146,6 +146,6 @@ public final class Pagination {
     
     private boolean isMaxRowCount(final ShardingSelectOptimizedStatement optimizedStatement) {
         return (!optimizedStatement.getGroupBy().getItems().isEmpty()
-                || !optimizedStatement.getAggregationSelectItems().isEmpty()) && !optimizedStatement.isSameGroupByAndOrderByItems();
+                || !optimizedStatement.getSelectItems().getAggregationSelectItems().isEmpty()) && !optimizedStatement.isSameGroupByAndOrderByItems();
     }
 }
